Обучит компонент карточки фильма принимать данные
