+++ src/const.js
@@ -1,10 +1,10 @@
 export const POSTERS = [
-  './public/images/posters/made-for-each-other.png',
-  './public/images/posters/popeye-meets-sinbad.png',
-  './public/images/posters/sagebrush-trail.jpg',
-  './public/images/posters/santa-claus-conquers-the-martians.jpg',
-  './public/images/posters/the-dance-of-life.jpg',
-  './public/images/posters/the-man-with-the-golden-arm.jpg',
+  './images/posters/made-for-each-other.png',
+  './images/posters/popeye-meets-sinbad.png',
+  './images/posters/sagebrush-trail.jpg',
+  './images/posters/santa-claus-conquers-the-martians.jpg',
+  './images/posters/the-dance-of-life.jpg',
+  './images/posters/the-man-with-the-golden-arm.jpg',
 ];
 
 export const EMOTIONS = {
