+++ src/view/popup/popup-comments-list-container-view.js
@@ -1,0 +1,25 @@
+import {createElement} from '../render.js';
+
+function createPopupCommentsListContainerTemplate() {
+  return (
+    '<ul class="film-details__comments-list"></ul>'
+  );
+}
+
+export default class PopupCommentsListContainerView {
+  getTemplate() {
+    return createPopupCommentsListContainerTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
