+++ src/view/popup/popup-container-view.js
@@ -1,0 +1,28 @@
+import {createElement} from '../render.js';
+
+function createPopupContainerTemplate() {
+  return (
+    `<section class="film-details">
+      <div class="film-details__inner">
+      </div>
+    </section>`
+  );
+}
+
+export default class PopupContainerView {
+  getTemplate() {
+    return createPopupContainerTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
